requirements.txt
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.20
langchain_openai==0.2.9
openai==1.55.0
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
streamlit==1.38.0
